Test port number string conversion to short instead of int
Port numbers have a small range and are typically expressed as short ints.
Therefore test the string conversion to short instead.

+++ src/blackmisc/dbusserver.cpp
@@ -367,7 +367,7 @@ namespace BlackMisc
         QString p = port.toLower().trimmed();
         if (!p.isEmpty())
         {
-            p.toInt(&ok);
+            p.toShort(&ok);
             if (!ok)
             {
                 p = ""; // was not a number
